+++ CHANGELOG.md
@@ -5,8 +5,21 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.11.2] - 2021-11-15
+### Fixed
+- Fixed spaces around brackets string (`[[string]]`) used as an index or table key (i.e. `[ [[string]] ]`) being removed, leading to a syntax error. ([#293](https://github.com/JohnnyMorganz/StyLua/issues/293))
+- Fixed incorrect shape calculation leading to arguments incorrectly expanding when under column width. ([#298](https://github.com/JohnnyMorganz/StyLua/issues/298))
+- Fixed incorrect shape calculation for singleline table at the column width boundary. ([#296](https://github.com/JohnnyMorganz/StyLua/issues/296))
+- Fixed IfExpression syntax containing extra/abnormal trailing whitespace when currently formatting as-is under the `luau` feature flag. ([#297](https://github.com/JohnnyMorganz/StyLua/issues/297))
+- Fixed newlines before arguments in a function call which is later formatted on a single line being preserved, leading to inconsistent formatting. ([#290](https://github.com/JohnnyMorganz/StyLua/issues/290))
+- Fixed odd formatting when returning multiple tables or functions only. ([#302](https://github.com/JohnnyMorganz/StyLua/issues/302))
+- Fixed comments within an index expression (`foo[index]`) incorrectly handled leading to malformed formatting. ([#304](https://github.com/JohnnyMorganz/StyLua/issues/304))
+
+## [0.11.1] - 2021-11-08
 ### Changed
 - Updated internal parser to fix parsing issues and update `luau` parsing. ([#229](https://github.com/JohnnyMorganz/StyLua/issues/229), [#231](https://github.com/JohnnyMorganz/StyLua/issues/231))
+- Default glob now matches `**/*.luau` (as well as `**/*.lua`) when the `luau` flag is enabled. ([#291](https://github.com/JohnnyMorganz/StyLua/issues/291))
 
 ### Fixed
 - Fixed indentation of type callback specifier parameters when parameters have leading comment trivia. ([#278](https://github.com/JohnnyMorganz/StyLua/issues/278))